decomp(laneAG2): deep body ports -> +100 (30,244); flip rate 71%
103 gained / 3 lost, unit-agnostic, re-verified by the coordinator over two
build legs plus a forced dtk re-split. ★An intermediate six-branch merge measured
+97, and that same +97 reproduced against TWO different baselines (30,101 and
30,144) -- the determinism control, and proof of independence from laneAE2's +43.
Flip rate 51 closed of 72 attempted = 71%, across 7 subagents.

The 3 LOST are BandCamShot EH funclets whose frame constants are BYTE-IDENTICAL
via /FAs -- objdiff's unnamed-symbol pairing ORDER shifted when the obj's COMDAT
set changed. Known funclet over-subscription, not recoverable from source.

★★CORRECTS A RULE WE USED ALL DAY: THE 78-96% FLIP BAND STRUCTURALLY EXCLUDED
THE CHEAPEST TIER. An `S=1` penalty (one differing immediate) ALWAYS scores
99.9x%, so it could never appear in that band -- and it is the LARGEST tier in
the binary at 42 functions. Tool: scripts/harvest/nearfree_tier_worklist.py.
★AND IT IS A LAYOUT TIER WEARING AN IMMEDIATE MASK: of 14 `S=1` targets, 0 map
mispairs and 0 reloc artifacts, but only 3 were literally one wrong constant --
9 were class/member-LAYOUT deltas, 2 stack-slot. Route it to LAYOUT work.

★★DECODER CORRECTIONS, one affecting honesty: `report.json` ROUNDS 99.953 UP TO
100.0, so strict counts can OVER-CREDIT. `S=5` is one differing OPCODE (never
the at_limit shape). Branch-target-only diffs are unscored under
functionRelocDiffs=none.

★NEW CAPABILITY: `cl.exe /d1reportSingleClassLayout<Class>` WORKS THROUGH WIBO
-- exact offsets, padding, vtables, `this` adjustors. Strictly better than the
stale `// 0xHEX` header comments that have misled several lanes; worth wiring
into the struct-info / stack-layout skills.

★`regswap => at_limit` FAILED TWICE MORE: a 23-instruction and an 18-instruction
regswap cascade each dissolved entirely from ONE statement-level change.

★TWO FIXERS REACHED OPPOSITE VERDICTS ON THE SAME `extsb.`/`cmplwi` INSTRUCTION
AND BOTH WERE PARTLY WRONG. Retail never calls strcpy; the temp must be a named
`unsigned int` READ TWICE IN SEPARATE STATEMENTS (declared width alone is
insufficient -- MSVC folds a single-use temp away). THREE loop shapes exist
(DELTA 127 / XOR 95 / UPDATE 49), only UPDATE is fixable, and the discriminating
test in circulation does NOT separate DELTA from UPDATE. ★Sweep priced at ZERO:
no remaining UPDATE site sits in a function >=90%. 271 sites, yield 2 -- the
site-count-is-not-defect-count trap caught BEFORE paying for it.

WORKTREE RECOVERY (done first): all 7 inspected, every uncommitted buffer
committed on its own branch and tagged before anything else was touched.
b3b +2, b3c +14, sysb +5, sysc +11 = exactly +32, fully additive, all four
self-reports accurate. ★The two branches with NO commits contributed 0 and their
buffers DID NOT BUILD (ambiguous stlpmtx_std::max; a Campaign.h map->hash_map
typedef breaking two consumers) -- **a dying lane's working tree is a LEAD, not
a patch.** Both leads were later salvaged productively.
★PROCESS RULE ADDED: a fixer that resumes AFTER integration commits BEHIND the
merge, and the merge commit gives no hint -- the lane nearly lost sysc's +3 that
way. Re-check `git merge-base --is-ancestor` for every branch immediately before
the final measurement.

POOL: 28 of the 439-member band closed, 372 remain; near-free tiers drained
harder (`S>=60` never-attempted 82 -> 45, `S in {1,2,5}` 44 -> 29). The earlier
"309" needs a third clustering axis (delta_shape) to reproduce and its "named"
set included `$`-bearing thunks -- a naive redo gives 327.
Also: kState_ChoosePartWarn does not exist on the disc, so the HX_NATIVE
workaround above it guards a path that never shipped.

SINGLE-OWNER HANDOFFS (reported, not applied): MAP -- HasCampaignKey@Campaign
points at GetCampaignKey's address (false 48.9%); GetHarmony@Stats mispair
(target is a float getter at +0xb8); SetFrameEx@HamCamShot possibly cross-class
paired. SPLITS -- BandCharDesc::operator== is a 4-byte target-carve truncation
with instructions 0-82 already byte-identical, should flip for free. Plus 7
layout and 3 stack-layout leads.

Whole-binary A/B: 30,144 -> 30,244, 103 gained / 3 lost (unit-agnostic).

+++ src/band3/meta_band/BandMachineMgr.cpp
@@ -329,7 +329,9 @@ DataNode BandMachineMgr::ForEachMachine(const DataArray *arr) {
     DataNode local(*var);
     std::vector<BandMachine *> machines;
     GetMachines(machines);
-    for (int i = 0; i < (int)machines.size(); i++) {
+    // Unsigned loop bound: retail's entry guard is `beq` (size == 0), not the
+    // `ble` a signed `(int)machines.size()` compare produces.
+    for (unsigned int i = 0; i < machines.size(); i++) {
         *var = machines[i];
         for (int j = 3; j < arr->Size(); j++) {
             arr->Command(j)->Execute();
